Move the validator and error-carrier tests to dockerclient
Same reason the fuzz targets moved: the code is in dockerclient now and
the tests were still in dockerapi, reaching it through the compatibility
aliases. dockerclient reported zero test functions, which is a fair thing
for a reader to hold against it, and a second implementation of the
interface inherited none of the coverage.

Seven tests move: ParseImageRef, CheckID and ContainerConfig.Validate
from the validation side, and the InvalidArgumentError, ConnectionError,
APIVersionError and ResponseError message tests from the error side.

The ones that stay in dockerapi are the ones genuinely about dockerapi:
that its endpoints apply the guards before sending a request, that the id
guard is on every container endpoint, the exec version gates, and
newStatusError, which builds a dockerclient error out of an HTTP response
and so belongs to the implementation rather than the interface.
